Implement basic auth for hg push command
