+++ scm-plugins/scm-hg-plugin/src/main/java/sonia/scm/repository/spi/HgPullCommand.java
@@ -47,7 +47,6 @@ import java.util.List;
 public class HgPullCommand extends AbstractHgPushOrPullCommand implements PullCommand {
 
   private static final Logger LOG = LoggerFactory.getLogger(HgPullCommand.class);
-  private static final String AUTH_SECTION = "auth";
   private final ScmEventBus eventBus;
   private final HgLazyChangesetResolver changesetResolver;
   private final HgRepositoryHookEventFactory eventFactory;
@@ -69,13 +68,14 @@ public class HgPullCommand extends AbstractHgPushOrPullCommand implements PullCo
   public PullResponse pull(PullCommandRequest request)
     throws IOException {
     String url = getRemoteUrl(request);
+    HgIniConfigurator iniConfigurator = new HgIniConfigurator(getContext());
 
     LOG.debug("pull changes from {} to {}", url, getContext().getScmRepository());
 
     List<Changeset> result;
 
     if (!Strings.isNullOrEmpty(request.getUsername()) && !Strings.isNullOrEmpty(request.getPassword())) {
-      addAuthenticationConfig(request, url);
+      iniConfigurator.addAuthenticationConfig(request, url);
     }
 
     try {
@@ -83,7 +83,7 @@ public class HgPullCommand extends AbstractHgPushOrPullCommand implements PullCo
     } catch (ExecutionException ex) {
       throw new ImportFailedException(ContextEntry.ContextBuilder.entity(getRepository()).build(), "could not execute pull command", ex);
     } finally {
-      removeAuthenticationConfig();
+      iniConfigurator.removeAuthenticationConfig();
     }
 
     firePostReceiveRepositoryHookEvent();
@@ -93,38 +93,5 @@ public class HgPullCommand extends AbstractHgPushOrPullCommand implements PullCo
 
   private void firePostReceiveRepositoryHookEvent() {
     eventBus.post(eventFactory.createEvent(context, changesetResolver));
-  }
-
-  public void addAuthenticationConfig(PullCommandRequest request, String url) throws IOException {
-    INIConfiguration ini = readIniConfiguration();
-    INISection authSection = ini.getSection(AUTH_SECTION);
-    if (authSection == null) {
-      authSection = new INISection(AUTH_SECTION);
-      ini.addSection(authSection);
-    }
-    URI parsedUrl = URI.create(url);
-    authSection.setParameter("import.prefix", parsedUrl.getHost());
-    authSection.setParameter("import.schemes", parsedUrl.getScheme());
-    authSection.setParameter("import.username", request.getUsername());
-    authSection.setParameter("import.password", request.getPassword());
-    writeIniConfiguration(ini);
-  }
-
-  public void removeAuthenticationConfig() throws IOException {
-    INIConfiguration ini = readIniConfiguration();
-    ini.removeSection(AUTH_SECTION);
-    writeIniConfiguration(ini);
-  }
-
-  public INIConfiguration readIniConfiguration() throws IOException {
-    return new INIConfigurationReader().read(getHgrcFile());
-  }
-
-  public void writeIniConfiguration(INIConfiguration ini) throws IOException {
-    new INIConfigurationWriter().write(ini, getHgrcFile());
-  }
-
-  public File getHgrcFile() {
-    return new File(getContext().getDirectory(), HgRepositoryHandler.PATH_HGRC);
   }
 }
